Updated pop_stack functions

+++ RAM.hpp
@@ -26,5 +26,5 @@ public:
 
     // Stack operations
     uint16_t push_stack(uint16_t sp, uint16_t data);
-    uint16_t pop_stack(uint16_t sp);
+    uint32_t pop_stack(uint16_t sp);
 };
